[Test] Clean up useless test for ascend scheduler
